requirements.txt
python-dotenv
PyTelegramBotAPI
texttable